core: move function_signature_t to type_traits
