+++ core/include/gap/core/type_traits.hpp
@@ -17,10 +17,80 @@ namespace gap::trait
 
     template< typename T >
     inline constexpr auto type = type_wrapper< T >{};
+
     template< typename T >
     constexpr auto typeof(T &&) {
         return type< T >;
     }
+
+
+    namespace detail
+    {
+        template< typename T, typename enable_t = void >
+        struct function_signature;
+
+        template< typename T >
+        struct function_signature< T, std::enable_if_t< std::is_class_v< std::remove_cvref_t< T > > > >
+            : function_signature< decltype(&std::remove_cvref_t< T >::operator()) >
+        {};
+
+        template< typename T, typename result_t, typename... args_t >
+        struct function_signature< result_t (T::*)(args_t...) > {
+            using type = result_t(args_t...);
+        };
+
+        template< typename T, typename result_t, typename... args_t >
+        struct function_signature< result_t (T::*)(args_t...) const > {
+            using type = result_t(args_t...);
+        };
+
+        template< typename result_t, typename... args_t >
+        struct function_signature< result_t(args_t...) > {
+            using type = result_t(args_t...);
+        };
+
+        template< typename result_t, typename... args_t >
+        struct function_signature< result_t (&)(args_t...) > {
+            using type = result_t(args_t...);
+        };
+
+        template< typename result_t, typename... args_t >
+         struct function_signature< result_t (*)(args_t...) > {
+            using type = result_t(args_t...);
+        };
+    } // namespace detail
+
+    template< typename T >
+    using function_signature_t = typename detail::function_signature< T >::type;
+
+    namespace detail
+    {
+        template< typename T >
+        struct drop_first_arg;
+
+        template< typename result_t, typename first_t, typename... args_t >
+        struct drop_first_arg< result_t(first_t, args_t...) > {
+            using type = result_t(args_t...);
+        };
+
+        struct dummy {};
+
+        template< typename T, typename result_t = dummy, typename enable_t = void >
+        struct recursive_function_signature;
+
+        template< typename T, typename result_t >
+        struct recursive_function_signature<
+            T,
+            result_t,
+            std::enable_if_t< std::is_class_v< std::remove_cvref_t< T > > > >
+        {
+            using type = typename drop_first_arg< function_signature_t<
+                decltype(&std::remove_cvref_t< T >::template operator()< result_t (*)(...) >) > >::type;
+        };
+    } // namespace detail
+
+    template< typename... T >
+    using recursive_function_signature_t = typename detail::recursive_function_signature< T... >::type;
 
     namespace
     {
@@ -30,9 +100,9 @@ namespace gap::trait
         template< typename T, typename U >
         constexpr bool check(T &&a, U &&b) {
             if constexpr (type< T > == type< U >) {
-                static_assert(typeof(a) == typeof(b));
+                return typeof(a) == typeof(b);
             } else {
-                static_assert(typeof(a) != typeof(b));
+                return typeof(a) != typeof(b);
             }
         }
